feat: add comments

+++ src/controllers/v1/user/delete_current_user.ts
@@ -6,17 +6,20 @@ import type { Request, Response } from 'express';
 
 const deleteCurrentUser = async (req: Request, res: Response) => {
   try {
-    await userModel.deleteOne({ _id: req.userId });
- 
     const blogs = await blogModel
       .find({ author: req.userId })
       .select('banner.publicId')
       .lean()
       .exec();
 
-    const publicIds = blogs.map((blog) => blog.banner.publicId);
-    await deleteMultipleFromCloudinary(publicIds);
-    await blogModel.deleteMany({ author: req.userId });
+    if (blogs.length > 0) {
+      const publicIds = blogs.map((blog) => blog.banner.publicId);
+      await deleteMultipleFromCloudinary(publicIds);
+      await blogModel.deleteMany({ author: req.userId });
+    }
+
+    
+    await userModel.deleteOne({ _id: req.userId });
 
     logger.info(`User deleted successfully`, { userId: req.userId });
